Get all processes and display them. On click display their stages.
Add process. Can edit name when not creating stage.

When creating stage can edit question, response type, and can add or subtract the amount of responses to question. On save, stage gets added to the end of the process's stageList.

Can edit stage order while creating process and can subtract stages.

When saving process, the values from the process get transferred to another interface inorder to be sent to the backend. When receiving processes from the backend, the type of interface get converted back.